% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimMobTools.R
\docType{package}
\name{ClimMobTools}
\alias{ClimMobTools}
\alias{ClimMobTools-package}
\title{API Client for the 'ClimMob' platform in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

API client for 'ClimMob', an open source software for crowdsourcing 
    citizen science in agriculture under the 'tricot' method <https://climmob.net/climmob3/>. 
    Developed by van Etten et al. (2019) <doi:10.1017/S0014479716000739>, it turns the 
    research paradigm on its head; instead of a few researchers designing complicated 
    trials to compare several technologies in search of the best solutions, 
    it enables many farmers to carry out reasonably simple experiments that 
    taken together can offer even more information. 'ClimMobTools' enables project 
    managers to deep explore and analyse their 'ClimMob' data in R.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrobioinfoservices/ClimMobTools}}
\item{Static documentation: 
 \url{https://agrobioinfoservices.github.io/ClimMobTools/}}
\item{Report bugs: 
 \url{https://github.com/agrobioinfoservices/ClimMobTools/issues}}
}
}
\author{
Kauê de Sousa and Jacob van Etten and Brandon Madriz
}
