% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourite.R
\name{favourite}
\alias{favourite}
\alias{favorite}
\alias{plot.cmb_fvrt}
\title{Favourability scores}
\usage{
favourite(data = NULL, items = NULL, input = NULL, reorder = TRUE)

\method{plot}{cmb_fvrt}(x, ...)
}
\arguments{
\item{data}{a data frame with columns specified by items and input values}

\item{items}{a data frame or index of \code{data} for the column(s) containing the item names}

\item{input}{a data frame or index of \code{data} for the column(s) containing the ranked values}

\item{reorder}{logical, if items should be reordered from higher favourability score to least favourability score}

\item{x}{an object of class 'cmb_fvrt' for the plotting method. Generates a 'ggplot' object that can be passed to any ggplot2 method}

\item{...}{additional arguments passed to methods. See details}
}
\value{
a data.frame with descriptive statistics
\item{N}{number of times the given item was evaluated}
\item{best}{relative number of times (in percentage) the given item was ranked as first}
\item{worst}{relative number of times (in percentage) the given item was ranked as worst}
\item{wins}{relative number of times (in percentage) the given item wins against the others}
\item{fav_score}{the favourability score, which is the difference between best and worst performance}
}
\description{
Compute favourability scores
}
\examples{

# beans data where each observer compares 3 varieties 
# randomly distributed 
 
library("PlackettLuce")
data("beans", package = "PlackettLuce")

favourite(data = beans,
          items = c(1:3),
          input = c(4:5))

# ...........................................
# ...........................................


# rankings with five items in a ClimMob project

items <- randomise(5, 5, 5, c("green","blue","red","white","yellow"))

input <- as.data.frame(matrix(NA, nrow = 5, ncol = 5))
names(input) <- paste0("position_item_",LETTERS[1:5])

for(s in 1:5) {
  input[s,] <- sample(1:5)
}

fav <- favourite(items = items, 
                 input = input)
                 
}
