% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrialMetadata.R
\name{getTrialMetadata}
\alias{getTrialMetadata}
\title{Extract trial metadata}
\usage{
getTrialMetadata(x)
}
\arguments{
\item{x}{a list with the raw ClimMob data}
}
\value{
a list with the trial metadata
}
\description{
Metadata for the export functions on data publication
}
