% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimMobTools.R
\docType{package}
\name{ClimMobTools}
\alias{ClimMobTools}
\alias{ClimMobTools-package}
\title{API Client for the 'ClimMob' platform in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

API client for 'ClimMob', an open source software for decentralized large-N trials with the 'tricot' approach \url{https://climmob.net/}. Developed by van Etten et al. (2019) \doi{10.1017/S0014479716000739}, it turns the research paradigm on its head; instead of a few researchers designing complicated trials to compare several technologies in search of the best solutions for the target environment, it enables many participants to carry out reasonably simple experiments that taken together can offer even more information. 'ClimMobTools' enables project managers to deep explore and analyse their 'ClimMob' data in R.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrdatasci/ClimMobTools}}
\item{Static documentation: 
 \url{https://agrdatasci.github.io/ClimMobTools/}}
\item{Report bugs: 
 \url{https://github.com/agrdatasci/ClimMobTools/issues}}
\item{ClimMob Platform: 
 \url{https://climmob.net}}
\item{The tricot user guide: 
 \url{https://hdl.handle.net/10568/109942}}
}
}
\author{
Kauê de Sousa and Jacob van Etten and Brandon Madriz
}
